\name{catage}
\alias{catage}
\docType{data}
\title{Catch numbers and weight at age for an Icelandic stock.
}
\description{
This data set gives catch@age ('ObsCno') and catch-weights@age ('CatchWts') for a fish stock in
a 5 column dataframe, indexed by year, age and year class for the period 1980--2009.
}

\usage{catage}
\format{A dataframe containing 360 observations of 5 variables.}
\source{NWWG2010}
\references{NWWG2010
}
\keyword{datasets}
