\name{hafroAssmt-package}
\alias{hafroAssmt-package}
\alias{hafroAssmt}
\docType{package}
\title{Hafro Assessment Toolbox}
\description{
  Functions used in fisheries stock assessment at Hafro, the Marine
  Research Institute in Iceland.
}
\details{
  \emph{Plot functions:}
  \tabular{ll}{
    \code{\link{labit}} \tab create label\cr
  }
}
\note{
  \code{browseVignettes()} shows the vignettes, found in the
  \file{hafroAssmt/doc} directory.
}
\author{Arni Magnusson and Hoskuldur Bjornsson with contributions from
  Sigurdur Thor Jonsson.}
\references{
  \url{http://hafroassmt.r-forge.r-project.org}.
}
% Programming
\keyword{utilities}
% Statistics
\keyword{models}
